+++ scripts/runprog_examples.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+
+./runprog.sh -v
+./runprog.sh -- runprog.sh -h
+./runprog.sh -- /usr/bin/ls